Apple TV: support Siri remote pan gesture (#295)
After this change, the developer can call a method that will turn on a gesture handler to detect pan events from the Siri remote touchscreen, and surface them in the JS layer. Thanks to @respectTheCode for the excellent module https://github.com/respectTheCode/react-native-tvos-controller , which is the basis for this feature. This PR fixes #16 .

- `TVMenuControl` is now named `TVEventControl`, and new methods are added to enable and disable pan gesture detection
- When the feature is enabled, `TVEventHandler` will surface a new type of event with eventType="pan", and a new "body" object with details of the pan gesture event (position, velocity, and state value (begun, ended, changed)).
- Example code added to `TVEventHandlerExample` in the RNTester app.
- README.md updated with details on this feature.