Calll the ShadowNode::cloneMultiple callback for each cloned node (#54431)
Summary:
Pull Request resolved: https://github.com/facebook/react-native/pull/54431

The original intention by the cloneMultiple method was to give the possibility of defining the cloning method only for the nodes definied by the list of families. But this way it is impossible to for example set the value of `runtimeShadowNodeReference` in the fragment, so instead we now pass the responsibility of cloning to the user for all the cloned nodes.

# Changelog

[General] [Changed] - cloneMultiple now invokes the callback for every cloned node, instead of doing that only for the nodes in `familiesToUpdate`

Reviewed By: javache

Differential Revision: D85852143

fbshipit-source-id: 9cf0540f546f5f070a918b8a4434ad40f78d91f6